third_party/WebKit/LayoutTests/media/track/track-webvtt-tc028-unsupported-markup.html

<!DOCTYPE html>
<title>Tests that unsupported markup is properly ignored.</title>
<script src="track-helpers.js"></script>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
// TODO(srirama.m): Rewrite the test in a better way.
// See https://codereview.chromium.org/2030383002/ for details.
check_cues_from_track("captions-webvtt/tc028-unsupported-markup.vtt", function(track) {
    var expected = [
        {
            innerHTML: "Bear is Coming!!!!!\nAnd what kind of a bear it is - just have look.",
            text: "<h1>Bear is Coming!!!!!</h1>\n<p>And what kind of a bear it is - just have <a href=\"webpage.html\">look</a>.</p>"
        },
        {
            innerHTML: "\n  I said Bear is coming!!!!\n  I said Bear is still coming!!!!\n",
            text: "<ul>\n  <li>I said Bear is coming!!!!</li>\n  <li>I said Bear is still coming!!!!</li>\n</ul>"
        },
        {
            innerHTML: "\n  I said Bear is coming now!!!!\n  \n  \n",
            text: "<ol>\n  <li>I said Bear is coming now!!!!</li>\n  <li><img src=\"bear.png\" alt=\"mighty bear\"></li>\n  <li><video src=\"bear_ad.webm\" controls></video></li>\n</ol>"
        }
    ];

    var cues = track.cues;
    assert_equals(cues.length, expected.length);
    for (var i = 0; i < cues.length; i++) {
        assert_equals(cues[i].getCueAsHTML().textContent, expected[i].innerHTML);
        assert_equals(cues[i].text, expected[i].text);
    }
});
</script>
